Create PTE.tex file to digitise a book
This is a book written by professors at Department of Physics, University of Rajasthan, Jaipur. It is for the development of physics through experiments in different places. Some part of book has been typed in latex to make it's digital version. Above text is not free from copyright. If you are copying from here, you are doing on your own risk.